requirements.txt
pdfminer.six>=20221105
PyMuPDF>=1.20.0
Pillow>=9.3.0
PyYAML>=6.0 